src/modules/users/services/SendForgotPasswordMailService.ts
import { injectable, inject } from 'tsyringe';

import AppError from '@shared/errors/AppError';
import IEmailProvider from '@shared/container/providers/EmailProvider/models/IEmailProvider';
import IUserTokensRepository from '../repositories/IUserTokensRepository';
import IUsersRepository from '../repositories/iUserRepository';

interface IRequest {
    email: string;
}

@injectable()
class SendForgotPasswordMailService {
    constructor(
        @inject('UsersRepository')
        private usersRepository: IUsersRepository,

        @inject('EmailProvider')
        private emailProvider: IEmailProvider,

        @inject('UserTokensRepository')
        private userTokensRepository: IUserTokensRepository,
    ) {}

    public async execute({ email }: IRequest): Promise<void> {
        const user = await this.usersRepository.findByEmail(email);

        if (!user) {
            throw new AppError('O usuário não existe');
        }

        await this.userTokensRepository.generate(user.id);

        this.emailProvider.sendMail(email, 'Pedido recebido');
    }
}
export default SendForgotPasswordMailService;
